am4-rust/scripts/requirements.txt
duckdb>=0.9.2
psycopg2-binary>=2.9.9
pandas>=2.1.0
pyarrow>=14.0.1 